models/Grok/app4/backend/requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
Flask-Bcrypt==1.0.1
PyJWT==2.4.0
